requirements.txt
streamlit
opencv-python
numpy
Pillow
sqlalchemy
bcrypt
deepface

